TRPMutableCollectionViewCell has some new functionality, it now tells the containing view when it’s text field is finished being edited, so that the owning view can update the triplistitem and trip model accordingly, it might have to create a new trip model, or it might have to update an existing trip models location (I’m not sure if that part works correctly yet) TRPMutableCollectionViewCell also manages CLLocation functionality too, if no location has been manually entered, it displays a button in the middle of itself to get the current location and set that as the title
-Needed Functionality:
There is currently no way to delete a trip from the collection view, I’m not sure how we want to have the user do that.
When triplistitem is inserted, it is inserted at  index 0, so it shows up at the top of the collection view, but when a trip model is created for that cell, the trip model has to be saved and I reload all of the trips after this point to make sure that everything is set correctly, this operation shuffles the order of the trips in the collection view, which is destroying the UX of this interaction.